requirements.txt
fastapi
uvicorn
python-dotenv
fpdf
mysql-connector-python
pydantic
twilio
razorpay
databases
asyncmy
requests
